requirements.txt
py-cord
black
python-dotenv
psutil